multigpu version first version
