+++ HDGan/HDGan.py
@@ -29,7 +29,67 @@ def to_img_dict(*inputs):
     if len(inputs[2].size()) != 1: 
         res['output_256'] = inputs[2]
     mean_var = (inputs[3], inputs[4])
+    
     return res, mean_var
+
+def get_KL_Loss(mu, logvar):
+    # import pdb; pdb.set_trace()
+    kld = mu.pow(2).add(logvar.mul(2).exp()).add(-1).mul(0.5).add(logvar.mul(-1))
+    kl_loss = torch.mean(kld)
+    return kl_loss
+
+
+def compute_d_pair_loss(real_logit, wrong_logit, fake_logit, real_labels, fake_labels):
+
+    criterion = nn.MSELoss()
+    real_d_loss = criterion(real_logit, real_labels)
+    wrong_d_loss = criterion(wrong_logit, fake_labels)
+    fake_d_loss = criterion(fake_logit, fake_labels)
+
+    discriminator_loss =\
+        real_d_loss + (wrong_d_loss + fake_d_loss) / 2.
+    return discriminator_loss
+
+
+def compute_d_img_loss(wrong_img_logit, real_img_logit, fake_img_logit, real_labels, fake_labels):
+
+    criterion = nn.MSELoss()
+    wrong_d_loss = criterion(wrong_img_logit, real_labels)
+    real_d_loss  = criterion(real_img_logit, real_labels)
+    fake_d_loss  = criterion(fake_img_logit, fake_labels)
+
+    return fake_d_loss + (wrong_d_loss+real_d_loss) / 2
+
+def compute_g_loss(fake_logit, real_labels):
+    criterion = nn.MSELoss()
+    generator_loss = criterion(fake_logit, real_labels)
+    return generator_loss
+
+def plot_imgs(samples, epoch, typ, name, path='', model_name=None):
+    tmpX = save_images(samples, save=not path == '', save_path=os.path.join(
+        path, '{}_epoch{}_{}.png'.format(name, epoch, typ)), dim_ordering='th')
+    plot_img(X=tmpX, win='{}_{}.png'.format(name, typ), env=model_name)
+
+
+def load_partial_state_dict(model, state_dict):
+
+    own_state = model.state_dict()
+    for name, param in state_dict.items():
+        if name not in own_state:
+            raise KeyError('unexpected key "{}" in state_dict'
+                           .format(name))
+        if isinstance(param, Parameter):
+            # backwards compatibility for serialized parameters
+            param = param.data
+        try:
+            own_state[name].copy_(param)
+        except:
+            print('While copying the parameter named {}, whose dimensions in the model are'
+                  ' {} and whose dimensions in the checkpoint are {}, ...'.format(
+                      name, own_state[name].size(), param.size()))
+            raise
+    print('>> load partial state dict: {} initialized'.format(len(state_dict)))
+
 
 def train_gans(dataset, model_root, model_name, netG, netD, args):
     """
@@ -47,15 +107,15 @@ def train_gans(dataset, model_root, model_name, netG, netD, args):
         Descriminator
     """
 
-    img_loss_ratio = 1.0   # the weight of img_loss.
     d_lr = args.d_lr
     g_lr = args.g_lr
     tot_epoch = args.maxepoch
 
     ''' get train and test data sampler '''
-    train_sampler = dataset[0].next_batch
-    test_sampler = dataset[1].next_batch
-    updates_per_epoch = int(dataset[0]._num_examples / args.batch_size)
+    train_sampler = iter(dataset[0]).next
+    test_sampler = iter(dataset[1]).next
+
+    updates_per_epoch = int(len(dataset[0]) / args.batch_size)
 
     ''' configure optimizer '''
     optimizerD = optim.Adam(netD.parameters(), lr=d_lr, betas=(0.5, 0.999))
@@ -96,6 +156,8 @@ def train_gans(dataset, model_root, model_name, netG, netD, args):
     g_loss_plot = plot_scalar(
         name="g_loss", env=model_name, rate=args.display_freq)
     lr_plot = plot_scalar(name="lr", env=model_name, rate=args.display_freq)
+    kl_loss_plot = plot_scalar(name="kl_loss", env=model_name, rate=args.display_freq)
+    
     all_keys = ["output_64", "output_128", "output_256"]
     g_plot_dict, d_plot_dict = {}, {}
     for this_key in all_keys:
@@ -106,14 +168,23 @@ def train_gans(dataset, model_root, model_name, netG, netD, args):
 
     #--------Generator niose placeholder used for testing------------#
     z = torch.FloatTensor(args.batch_size, args.noise_dim).normal_(0, 1)
-    z = to_device(z, netG.device_id, requires_grad=False)
-    fixed_images, _, fixed_embeddings, _ = test_sampler(n_embed=1)
-    fixed_embeddings = to_device(
-        fixed_embeddings, netG.device_id, volatile=True)
+    z = to_device(z)
+    fixed_images, _, fixed_embeddings, _, _ = test_sampler()
+    fixed_embeddings = to_device(fixed_embeddings)
     fixed_z_data = [torch.FloatTensor(args.batch_size, args.noise_dim).normal_(
         0, 1) for _ in range(args.test_sample_num)]
-    fixed_z_list = [to_device(a, netG.device_id, volatile=True)
-                    for a in fixed_z_data]
+    fixed_z_list = [to_device(a) for a in fixed_z_data]
+    
+    REAL_global_LABELS = Variable(torch.FloatTensor(args.batch_size, 1).fill_(1)).cuda()
+    FAKE_global_LABELS = Variable(torch.FloatTensor(args.batch_size, 1).fill_(0)).cuda()
+    REAL_local_LABELS = Variable(torch.FloatTensor(args.batch_size, 1, 4, 4).fill_(1)).cuda()
+    FAKE_local_LABELS = Variable(torch.FloatTensor(args.batch_size, 1, 4, 4).fill_(0)).cuda()
+
+    def get_labels(logit):
+        if logit.size(-1) == 1: 
+            return REAL_global_LABELS.view_as(logit), FAKE_global_LABELS.view_as(logit)
+        else:
+            return REAL_local_LABELS.view_as(logit), FAKE_local_LABELS.view_as(logit)
 
     #--------Start training------------#
     for epoch in range(start_epoch, tot_epoch):
@@ -126,82 +197,89 @@ def train_gans(dataset, model_root, model_name, netG, netD, args):
             set_lr(optimizerD, d_lr)
             set_lr(optimizerG, g_lr)
 
+        train_sampler = iter(dataset[0]).next # reset
+        test_sampler = iter(dataset[1]).next
+        netG.train()
+        netD.train()
         for it in range(updates_per_epoch):
-            netG.train()
             ncritic = args.ncritic
 
             for _ in range(ncritic):
                 ''' Sample data '''
-                images, wrong_images, np_embeddings, _ = train_sampler()
-                embeddings = to_device(np_embeddings, netD.device_id, requires_grad=False)
+                try:
+                    images, wrong_images, np_embeddings, _, _ = train_sampler.next()
+                except:
+                    train_sampler = iter(dataset[0]) # reset
+                    images, wrong_images, np_embeddings, _, _ = train_sampler.next()
+                    
+                embeddings = to_device(np_embeddings)
                 z.data.normal_(0, 1)
 
                 ''' update D '''
-                for p in netD.parameters():
-                    p.requires_grad = True
+                for p in netD.parameters(): p.requires_grad = True
                 netD.zero_grad()
 
-                g_emb = Variable(embeddings.data, volatile=True)
-                g_z = Variable(z.data, volatile=True)
-                fake_images, _ = netG(g_emb, g_z)
+                fake_images, mean_var = to_img_dict(netG(embeddings, z))
 
                 discriminator_loss = 0
                 ''' iterate over image of different sizes.'''
                 for key, _ in fake_images.items():
-                    this_img = to_device(images[key], netD.device_id)
-                    this_wrong = to_device(wrong_images[key], netD.device_id)
+                    this_img = to_device(images[key])
+                    this_wrong = to_device(wrong_images[key])
                     this_fake = Variable(fake_images[key].data)
-            
-                    real_dict = netD(this_img, embeddings)
-                    wrong_dict = netD(this_wrong, embeddings)
-                    fake_dict = netD(this_fake, embeddings)
-                    real_logit,  real_img_logit_local = real_dict['pair_disc'],  real_dict['local_img_disc']
-                    wrong_logit, wrong_img_logit_local = wrong_dict[
-                        'pair_disc'], wrong_dict['local_img_disc']
-                    fake_logit,  fake_img_logit_local = fake_dict['pair_disc'],  fake_dict['local_img_disc']
 
-                    ''' compute pair loss '''
-                    discriminator_loss += compute_d_pair_loss(
-                        real_logit, wrong_logit, fake_logit)
-                    ''' compute image loss '''
-                    img_loss = compute_d_img_loss(
-                        wrong_img_logit_local,  real_img_logit_local,   fake_img_logit_local,  prob=0.5)
+                    real_logit,  real_img_logit_local = netD(this_img, embeddings)
+                    wrong_logit, wrong_img_logit_local = netD(this_wrong, embeddings)
+                    fake_logit,  fake_img_logit_local = netD(this_fake, embeddings)
 
-                    discriminator_loss += img_loss_ratio * img_loss
+                    ''' compute disc pair loss '''
+                    real_labels, fake_labels = get_labels(real_logit)
+                    pair_loss =  compute_d_pair_loss(real_logit, wrong_logit, fake_logit, real_labels, fake_labels)
+
+                    ''' compute disc image loss '''
+                    real_labels, fake_labels = get_labels(real_img_logit_local)
+                    img_loss = compute_d_img_loss(wrong_img_logit_local, real_img_logit_local, fake_img_logit_local, real_labels, fake_labels)
+
+                    discriminator_loss += (pair_loss + img_loss)
+
                     d_plot_dict[key].plot(img_loss.cpu().data.numpy().mean())
 
                 d_loss_val = discriminator_loss.cpu().data.numpy().mean()
                 discriminator_loss.backward()
+
                 optimizerD.step()
                 netD.zero_grad()
+
                 d_loss_plot.plot(d_loss_val)
 
             ''' update G '''
-            for p in netD.parameters():
-                p.requires_grad = False  # to avoid computation
+            for p in netD.parameters(): p.requires_grad = False  # to avoid computation
             netG.zero_grad()
 
             # TODO Test if we do need to sample again
-            z.data.normal_(0, 1)  # resample random noises
-            fake_images, kl_loss = netG(embeddings, z)
+            # z.data.normal_(0, 1)  # resample random noises
+            # fake_images, kl_loss = netG(embeddings, z)
 
             loss_val = 0
+            kl_loss = get_KL_Loss(mean_var[0], mean_var[1])
+            kl_loss_val = kl_loss.cpu().data.numpy().mean()
             generator_loss = args.KL_COE*kl_loss
-
+            kl_loss_plot.plot(kl_loss_val)
             #---- iterate over image of different sizes ----#
+            '''Compute gen loss'''
             for key, _ in fake_images.items():
                 this_fake = fake_images[key]
-                fake_dict = netD(this_fake,  embeddings)
-                fake_pair_logit, fake_img_logit_local = fake_dict[
-                    'pair_disc'], fake_dict['local_img_disc']
+                fake_pair_logit, fake_img_logit_local = netD(this_fake, embeddings)
 
                 # -- compute pair loss ---
-                generator_loss += compute_g_loss(fake_pair_logit)
-                # -- compute image loss ---
-                img_loss_ = compute_g_loss(fake_img_logit_local)
+                real_labels, _ = get_labels(fake_pair_logit)
+                generator_loss += compute_g_loss(fake_pair_logit, real_labels)
 
-                generator_loss += img_loss_ * img_loss_ratio
-                g_plot_dict[key].plot(img_loss_.cpu().data.numpy().mean())
+                # -- compute image loss ---
+                real_labels, _ = get_labels(fake_img_logit_local)
+                img_loss = compute_g_loss(fake_img_logit_local, real_labels)
+                generator_loss += img_loss
+                g_plot_dict[key].plot(img_loss.cpu().data.numpy().mean())
 
             generator_loss.backward()
             g_loss_val = generator_loss.cpu().data.numpy().mean()
@@ -214,7 +292,7 @@ def train_gans(dataset, model_root, model_name, netG, netD, args):
             # --- visualize train samples----
             if it % args.verbose_per_iter == 0:
                 for k, sample in fake_images.items():
-                    plot_imgs([images[k], sample.cpu().data.numpy()],
+                    plot_imgs([images[k].numpy(), sample.cpu().data.numpy()],
                               epoch, k, 'train_images', model_name=model_name)
                 print('[epoch %d/%d iter %d/%d]: lr = %.6f g_loss = %.5f d_loss= %.5f' %
                       (epoch, tot_epoch, it, updates_per_epoch, g_lr, g_loss_val, d_loss_val))
@@ -227,7 +305,7 @@ def train_gans(dataset, model_root, model_name, netG, netD, args):
             if idx_test == 0:
                 test_images, test_embeddings = fixed_images, fixed_embeddings
             else:
-                test_images, _, test_embeddings, _ = test_sampler(n_embed=1)
+                test_images, _, test_embeddings, _ = test_sampler()
                 test_embeddings = to_device(
                     test_embeddings, netG.device_id, volatile=True)
                 testing_z = Variable(z.data, volatile=True)
@@ -237,7 +315,8 @@ def train_gans(dataset, model_root, model_name, netG, netD, args):
                     testing_z = fixed_z_list[t]
                 else:
                     testing_z.data.normal_(0, 1)
-                samples, _ = netG(test_embeddings, testing_z)
+                samples, _ = to_img_dict(netG(test_embeddings, testing_z))
+                 
                 if idx_test == 0 and t == 0:
                     for k in samples.keys():
                         #  +1 to make space for real image
@@ -262,16 +341,18 @@ def train_gans(dataset, model_root, model_name, netG, netD, args):
         end_timer = time.time() - start_timer
         # visualize testing samples
         for typ, v in vis_samples.items():
-            plot_imgs(v, epoch, typ, 'test_samples',
-                      path=model_folder, model_name=model_name)
+            plot_imgs(v, epoch, typ, 'test_samples', path=model_folder, model_name=model_name)
 
         ''' save weights '''
         if epoch % args.save_freq == 0:
-            netD = netD.cpu()
-            netG = netG.cpu()
-            torch.save(netD.state_dict(), os.path.join(
+            if 'DataParallel' in str(type(netD)):
+                netD_ = netD.module
+                netG_ = netG.module
+            netD_ = netD_.cpu()
+            netG_ = netG_.cpu()
+            torch.save(netD_.state_dict(), os.path.join(
                 model_folder, 'D_epoch{}.pth'.format(epoch)))
-            torch.save(netG.state_dict(), os.path.join(
+            torch.save(netG_.state_dict(), os.path.join(
                 model_folder, 'G_epoch{}.pth'.format(epoch)))
             print('save weights at {}'.format(model_folder))
 
@@ -281,62 +362,3 @@ def train_gans(dataset, model_root, model_name, netG, netD, args):
             'epoch {}/{} finished [time = {}s] ...'.format(epoch, tot_epoch, end_timer))
 
 
-def get_KL_Loss(mu, logvar):
-    # import pdb; pdb.set_trace()
-    kld = mu.pow(2).add(logvar.mul(2).exp()).add(-1).mul(0.5).add(logvar.mul(-1))
-    kl_loss = torch.mean(kld)
-    return kl_loss
-
-def compute_d_pair_loss(real_logit, wrong_logit, fake_logit,  real_labels, fake_labels):
-
-    criterion = nn.MSELoss()
-   
-    real_d_loss = criterion(real_logit, real_labels)
-    wrong_d_loss = criterion(wrong_logit, fake_labels)
-    fake_d_loss = criterion(fake_logit, fake_labels)
-
-    discriminator_loss =\
-        real_d_loss + (wrong_d_loss + fake_d_loss) / 2.
-    return discriminator_loss
-
-def compute_d_img_loss(wrong_img_logit, real_img_logit, fake_img_logit, real_labels, fake_labels):
-
-    criterion = nn.MSELoss()
-    wrong_d_loss = criterion(wrong_img_logit, real_labels)
-    real_d_loss  = criterion(real_img_logit, real_labels)
-    fake_d_loss  = criterion(fake_img_logit, fake_labels)
-
-    return fake_d_loss + (wrong_d_loss+real_d_loss) / 2
-
-
-def compute_g_loss(fake_logit, real_labels):
-
-    criterion = nn.MSELoss()
-    generator_loss = criterion(fake_logit, real_labels)
-    return generator_loss
-
-
-def plot_imgs(samples, epoch, typ, name, path='', model_name=None):
-    tmpX = save_images(samples, save=not path == '', save_path=os.path.join(
-        path, '{}_epoch{}_{}.png'.format(name, epoch, typ)), dim_ordering='th')
-    plot_img(X=tmpX, win='{}_{}.png'.format(name, typ), env=model_name)
-
-
-def load_partial_state_dict(model, state_dict):
-
-    own_state = model.state_dict()
-    for name, param in state_dict.items():
-        if name not in own_state:
-            raise KeyError('unexpected key "{}" in state_dict'
-                           .format(name))
-        if isinstance(param, Parameter):
-            # backwards compatibility for serialized parameters
-            param = param.data
-        try:
-            own_state[name].copy_(param)
-        except:
-            print('While copying the parameter named {}, whose dimensions in the model are'
-                  ' {} and whose dimensions in the checkpoint are {}, ...'.format(
-                      name, own_state[name].size(), param.size()))
-            raise
-    print('>> load partial state dict: {} initialized'.format(len(state_dict)))
