+++ test/test_gan/test_coco.sh
@@ -1,0 +1,5 @@
+
+name='coco_256'
+CUDA_VISIBLE_DEVICES=${device} python test_worker.py --dataset coco --model_name ${name} --load_from_epoch 200 --test_sample_num 2 --batch_size 32
+ 
+
